+++ CHANGELOG.md
@@ -22,3 +22,5 @@ Date: 2016-07-14
 - Bugfix for ScreenScraper who likes to put "N/A" on ratings making it unparseable as a number
 - Better exception catching, now if something bad happens with an API it doesn't stop
 
+## Unreleased
+- Increased http timeout to 60s