+++ CHANGELOG.md
@@ -26,5 +26,7 @@ Date: 2016-07-14
 - Increased http timeout to 90s
 - Better exception management when TheGamesDB is down
 - The games db changed the turbografx cd alias
+- Enabled SSL for ScreenScraper (and got back my account)
+- Better rom name cleaning